examples/djpcms-site/requirements.txt
-e git+http://github.com/lsbardel/djpcms.git/#egg=djpcms
psycopg2
django-socialauth
python-openid>=2.0.0
oauth>=1.0.0
south
-e git+http://github.com/facebook/python-sdk.git#egg=facebook_python_sdk
hg+https://pylinkedin.googlecode.com/hg/pylinkedin>=0.1